chore: re-org dirs
