+++ src/plugins/walletConnectToDapps/components/DAppInfo.tsx
