+++ src/plugins/walletConnectToDapps/components/header/ConnectDapp.tsx
@@ -2,7 +2,7 @@ import { ChevronRightIcon } from '@chakra-ui/icons'
 import type { ButtonProps } from '@chakra-ui/react'
 import { Button, useDisclosure } from '@chakra-ui/react'
 import { ConnectModal } from 'plugins/walletConnectToDapps/components/modals/connect/Connect'
-import { useWalletConnectV2 } from 'plugins/walletConnectToDapps/v2/WalletConnectV2Provider'
+import { useWalletConnectV2 } from 'plugins/walletConnectToDapps/WalletConnectV2Provider'
 import { useMemo } from 'react'
 import { useTranslate } from 'react-polyglot'
 import { WalletConnectIcon } from 'components/Icons/WalletConnectIcon'
