+++ src/plugins/walletConnectToDapps/components/modals/EIP155SignTypedDataConfirmation.tsx
@@ -2,9 +2,9 @@ import { Button, VStack } from '@chakra-ui/react'
 import { AddressSummaryCard } from 'plugins/walletConnectToDapps/components/modals/AddressSummaryCard'
 import { ModalSection } from 'plugins/walletConnectToDapps/components/modals/ModalSection'
 import { TypedMessageInfo } from 'plugins/walletConnectToDapps/components/modals/TypedMessageInfo'
-import { useWalletConnectState } from 'plugins/walletConnectToDapps/v2/hooks/useWalletConnectState'
-import type { EthSignTypedDataCallRequest } from 'plugins/walletConnectToDapps/v2/types'
-import type { WalletConnectRequestModalProps } from 'plugins/walletConnectToDapps/v2/WalletConnectModalManager'
+import { useWalletConnectState } from 'plugins/walletConnectToDapps/hooks/useWalletConnectState'
+import type { EthSignTypedDataCallRequest } from 'plugins/walletConnectToDapps/types'
+import type { WalletConnectRequestModalProps } from 'plugins/walletConnectToDapps/WalletConnectModalManager'
 import type { FC } from 'react'
 import { useTranslate } from 'react-polyglot'
 import { FoxIcon } from 'components/Icons/FoxIcon'
