+++ src/plugins/walletConnectToDapps/components/modals/SessionProposal.tsx
@@ -4,11 +4,11 @@ import { fromAccountId } from '@shapeshiftoss/caip'
 import type { HDWallet } from '@shapeshiftoss/hdwallet-core'
 import type { ProposalTypes, SessionTypes } from '@walletconnect/types'
 import { getSdkError } from '@walletconnect/utils'
+import { DAppInfo } from 'plugins/walletConnectToDapps/components/DAppInfo'
 import { ModalSection } from 'plugins/walletConnectToDapps/components/modals/ModalSection'
-import { DAppInfo } from 'plugins/walletConnectToDapps/v2/components/DAppInfo'
-import { Permissions } from 'plugins/walletConnectToDapps/v2/components/Permissions'
-import { WalletConnectActionType } from 'plugins/walletConnectToDapps/v2/types'
-import type { WalletConnectSessionModalProps } from 'plugins/walletConnectToDapps/v2/WalletConnectModalManager'
+import { Permissions } from 'plugins/walletConnectToDapps/components/Permissions'
+import { WalletConnectActionType } from 'plugins/walletConnectToDapps/types'
+import type { WalletConnectSessionModalProps } from 'plugins/walletConnectToDapps/WalletConnectModalManager'
 import type { FC } from 'react'
 import { useCallback, useState } from 'react'
 import { useTranslate } from 'react-polyglot'
