+++ src/plugins/walletConnectToDapps/hooks/useCallRequestEvmFees.ts
@@ -1,10 +1,10 @@
 import { fromAccountId } from '@shapeshiftoss/caip'
 import type { EvmBaseAdapter, EvmChainId } from '@shapeshiftoss/chain-adapters'
 import { FeeDataKey } from '@shapeshiftoss/chain-adapters'
+import { useWalletConnectState } from 'plugins/walletConnectToDapps/hooks/useWalletConnectState'
+import { assertIsTransactionParams } from 'plugins/walletConnectToDapps/typeGuards'
+import type { WalletConnectState } from 'plugins/walletConnectToDapps/types'
 import { getFeesForTx } from 'plugins/walletConnectToDapps/utils'
-import { useWalletConnectState } from 'plugins/walletConnectToDapps/v2/hooks/useWalletConnectState'
-import { assertIsTransactionParams } from 'plugins/walletConnectToDapps/v2/typeGuards'
-import type { WalletConnectState } from 'plugins/walletConnectToDapps/v2/types'
 import { useEffect, useMemo, useState } from 'react'
 import type { FeePrice } from 'components/Modals/Send/views/Confirm'
 import { getChainAdapterManager } from 'context/PluginProvider/chainAdapterSingleton'
